[media] s5p-mfc: split streaming status check per queue in VIDIOC_S_FMT
The vidioc_s_fmt() handler checks if streaming started for either the
OUTPUT or CAPTURE queue and returns -EBUSY if VIDIOC_S_FMT is called.

This is wrong since is valid to set the format of the CAPTURE buffer
queue even when streaming for the OUTPUT streaming has already started.

For example, GStreamer attempts to set the format for the CAPTURE queue
after playback started for the OUTPUT queue so this leads to following:

[  148.321594] s5p-mfc 11000000.codec: vidioc_s_fmt queue busy

Instead of checking for either of them, check only the status for the
corresponding OUTPUT or CAPTURE videobuf2 queue.

Signed-off-by: Javier Martinez Canillas <javier@osg.samsung.com>